(#28) fix: maxHeapMemoryUsage 변수 double -> int 변경

+++ server/src/main/java/ygo/traffic_hunter/core/dto/response/statistics/metric/StatisticsMetricMaxResponse.java
@@ -9,7 +9,7 @@ public record StatisticsMetricMaxResponse(
         double maxSystemCpuUsage,
         double maxProcessCpuUsage,
 
-        double maxHeapMemoryUsage,
+        int maxHeapMemoryUsage,
 
         int maxThreadCount,
         int maxPeakThreadCount,
